Steps to reproduce this bug:
1. Run tests in recorded mode with KEYVAULT_URI env var having a
   trailing slash. Then playback
2. In `testAuthentication` add a trailing slash to the value of
   `KEYVAULT_URI`, then run tests in recorded mode with KEYVAULT_URI env
   var having NO trailing slash. Then playback

Some of this could be our ID parsing logic which uses the URL! Really
I'm looking for guidance / recommendation - even if it means "ensure env
vars do not end in /" which I can document.

Left a bunch of console.log statements for diagnosis, they of course
should not end on master or anything.

Let me know if I can help